Permit root level output of CSS files.
