+++ lib/japr.rb
@@ -28,6 +28,7 @@ require 'japr/asset'
 require 'japr/converter'
 require 'japr/compressor'
 require 'japr/template'
+require 'japr/templates/template_helper'
 require 'japr/templates/javascript_tag_template'
 require 'japr/templates/css_tag_template'
 require 'japr/pipeline'
